Split large test files into partial classes
