+++ VirusTotalAnalyzer.Tests/VirusTotalClientTests.Submissions.cs
@@ -414,431 +414,4 @@ public partial class VirusTotalClientTests
         Assert.Equal("/api/v3/intelligence/retrohunt_jobs/rj1", handler.Request.RequestUri!.AbsolutePath);
     }
 
-    [Fact]
-    public async Task ListRetrohuntNotificationsAsync_PagesThroughResults()
-    {
-        var first = "{\"data\":[{\"id\":\"n1\",\"type\":\"retrohuntNotification\",\"data\":{\"attributes\":{\"job_id\":\"j1\"}}}],\"meta\":{\"cursor\":\"abc\"}}";
-        var second = "{\"data\":[{\"id\":\"n2\",\"type\":\"retrohuntNotification\",\"data\":{\"attributes\":{\"job_id\":\"j2\"}}}]}";
-        var handler = new QueueHandler(
-            new HttpResponseMessage(HttpStatusCode.OK) { Content = new StringContent(first, Encoding.UTF8, "application/json") },
-            new HttpResponseMessage(HttpStatusCode.OK) { Content = new StringContent(second, Encoding.UTF8, "application/json") });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var notifications = await client.ListRetrohuntNotificationsAsync(limit: 1);
-
-        Assert.Equal(2, notifications.Count);
-        Assert.Equal("n1", notifications[0].Id);
-        Assert.Equal("n2", notifications[1].Id);
-        Assert.Equal(2, handler.Requests.Count);
-        Assert.Contains("limit=1", handler.Requests[0].RequestUri!.Query);
-        Assert.Contains("cursor=abc", handler.Requests[1].RequestUri!.Query);
-    }
-
-    [Fact]
-    public async Task GetIpAddressReportAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"id\":\"1.1.1.1\",\"type\":\"ipAddress\",\"data\":{\"attributes\":{\"ip_address\":\"1.1.1.1\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var report = await client.GetIpAddressReportAsync("1.1.1.1");
-
-        Assert.NotNull(report);
-        Assert.Equal("1.1.1.1", report!.Id);
-        Assert.Equal(ResourceType.IpAddress, report.Type);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/ip_addresses/1.1.1.1", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetDomainReportAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"id\":\"example.com\",\"type\":\"domain\",\"data\":{\"attributes\":{\"domain\":\"example.com\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var report = await client.GetDomainReportAsync("example.com");
-
-        Assert.NotNull(report);
-        Assert.Equal("example.com", report!.Id);
-        Assert.Equal(ResourceType.Domain, report.Type);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/domains/example.com", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetAnalysisAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"id\":\"an1\",\"type\":\"analysis\",\"data\":{\"attributes\":{\"status\":\"queued\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var report = await client.GetAnalysisAsync("an1");
-
-        Assert.NotNull(report);
-        Assert.Equal("an1", report!.Id);
-        Assert.Equal(ResourceType.Analysis, report.Type);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/analyses/an1", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetAnalysisAsync_ThrowsApiException()
-    {
-        var errorJson = @"{""error"":{""code"":""NotFoundError"",""message"":""not found""}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.NotFound)
-        {
-            Content = new StringContent(errorJson, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var ex = await Assert.ThrowsAsync<ApiException>(() => client.GetAnalysisAsync("an1"));
-
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/analyses/an1", handler.Request!.RequestUri!.AbsolutePath);
-        Assert.Equal("NotFoundError", ex.Error?.Code);
-        Assert.Equal("not found", ex.Message);
-    }
-
-    [Fact]
-    public async Task GetGraphAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"id\":\"g1\",\"type\":\"graph\",\"data\":{\"attributes\":{\"name\":\"demo\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var graph = await client.GetGraphAsync("g1");
-
-        Assert.NotNull(graph);
-        Assert.Equal("g1", graph!.Id);
-        Assert.Equal(ResourceType.Graph, graph.Type);
-        Assert.Equal("demo", graph.Data.Attributes.Name);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/graphs/g1", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetCollectionAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"id\":\"c1\",\"type\":\"collection\",\"data\":{\"attributes\":{\"name\":\"demo\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var collection = await client.GetCollectionAsync("c1");
-
-        Assert.NotNull(collection);
-        Assert.Equal("c1", collection!.Id);
-        Assert.Equal(ResourceType.Collection, collection.Type);
-        Assert.Equal("demo", collection.Data.Attributes.Name);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/collections/c1", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetVotesAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"data\":[{\"id\":\"v1\",\"type\":\"vote\",\"data\":{\"attributes\":{\"date\":1,\"verdict\":\"malicious\"}}}]}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var votes = await client.GetVotesAsync(ResourceType.File, "abc");
-
-        Assert.NotNull(votes);
-        Assert.Single(votes!);
-        Assert.Equal("v1", votes[0].Id);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/votes", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetFeedAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"data\":[{\"id\":\"f1\",\"type\":\"file\"}]}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var feed = await client.GetFeedAsync(ResourceType.File);
-
-        Assert.NotNull(feed);
-        Assert.Single(feed!.Data);
-        Assert.Equal("f1", feed.Data[0].Id);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/feeds/files", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task GetRelationshipsAsync_DeserializesResponseAndUsesCorrectPath()
-    {
-        var json = "{\"data\":[{\"id\":\"r1\",\"type\":\"file\"}]}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var relationships = await client.GetRelationshipsAsync(ResourceType.File, "abc", "comments");
-
-        Assert.NotNull(relationships);
-        Assert.Single(relationships!.Data);
-        Assert.Equal("r1", relationships.Data[0].Id);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/relationships/comments", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task ReanalyzeFileAsync_UsesFilePath()
-    {
-        var json = "{\"id\":\"an\",\"type\":\"analysis\",\"data\":{\"attributes\":{\"status\":\"queued\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var report = await client.ReanalyzeFileAsync("abc");
-
-        Assert.NotNull(report);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/analyse", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task ReanalyzeUrlAsync_UsesUrlPath()
-    {
-        var json = "{\"id\":\"an\",\"type\":\"analysis\",\"data\":{\"attributes\":{\"status\":\"queued\"}}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var report = await client.ReanalyzeUrlAsync("def");
-
-        Assert.NotNull(report);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/urls/def/analyse", handler.Request!.RequestUri!.AbsolutePath);
-    }
-
-    [Fact]
-    public async Task ListRetrohuntJobsAsync_PagesThroughResults()
-    {
-        var first = "{\"data\":[{\"id\":\"j1\",\"type\":\"retrohuntJob\",\"data\":{\"attributes\":{\"status\":\"queued\"}}}],\"meta\":{\"cursor\":\"abc\"}}";
-        var second = "{\"data\":[{\"id\":\"j2\",\"type\":\"retrohuntJob\",\"data\":{\"attributes\":{\"status\":\"done\"}}}]}";
-        var handler = new QueueHandler(
-            new HttpResponseMessage(HttpStatusCode.OK)
-            {
-                Content = new StringContent(first, Encoding.UTF8, "application/json")
-            },
-            new HttpResponseMessage(HttpStatusCode.OK)
-            {
-                Content = new StringContent(second, Encoding.UTF8, "application/json")
-            });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var jobs = await client.ListRetrohuntJobsAsync(limit: 1);
-
-        Assert.Equal(2, jobs.Count);
-        Assert.Equal("j1", jobs[0].Id);
-        Assert.Equal("j2", jobs[1].Id);
-        Assert.Equal(2, handler.Requests.Count);
-        Assert.Contains("limit=1", handler.Requests[0].RequestUri!.Query);
-        Assert.Contains("cursor=abc", handler.Requests[1].RequestUri!.Query);
-    }
-
-    [Fact]
-    public async Task GetFileBehaviorAsync_DeserializesResponse()
-    {
-        var json = "{\"data\":[{\"id\":\"b1\",\"type\":\"analysis\",\"attributes\":{\"processes\":[{\"name\":\"proc1\"}]}}]}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var behavior = await client.GetFileBehaviorAsync("abc");
-
-        Assert.NotNull(behavior);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/behavior", handler.Request!.RequestUri!.AbsolutePath);
-        Assert.Equal("b1", behavior!.Data[0].Id);
-        Assert.Equal("proc1", behavior.Data[0].Attributes.Processes[0].Name);
-    }
-
-    [Fact]
-    public async Task GetFileBehaviorAsync_ThrowsApiException()
-    {
-        var errorJson = @"{""error"":{""code"":""NotFoundError"",""message"":""not found""}}";
-        var response = new HttpResponseMessage(HttpStatusCode.NotFound)
-        {
-            Content = new StringContent(errorJson, Encoding.UTF8, "application/json")
-        };
-        var handler = new SingleResponseHandler(response);
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var ex = await Assert.ThrowsAsync<ApiException>(() => client.GetFileBehaviorAsync("abc"));
-        Assert.Equal("not found", ex.Message);
-        Assert.Equal("NotFoundError", ex.Error?.Code);
-    }
-
-    [Fact]
-    public async Task GetFileBehaviorSummaryAsync_DeserializesResponse()
-    {
-        var json = "{\"data\":{\"tags\":[\"network\"]}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var summary = await client.GetFileBehaviorSummaryAsync("abc");
-
-        Assert.NotNull(summary);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/behavior_summary", handler.Request!.RequestUri!.AbsolutePath);
-        Assert.Contains("network", summary!.Data.Tags);
-    }
-
-    [Fact]
-    public async Task GetFileBehaviorSummaryAsync_ThrowsApiException()
-    {
-        var errorJson = @"{""error"":{""code"":""NotFoundError"",""message"":""not found""}}";
-        var response = new HttpResponseMessage(HttpStatusCode.NotFound)
-        {
-            Content = new StringContent(errorJson, Encoding.UTF8, "application/json")
-        };
-        var handler = new SingleResponseHandler(response);
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        await Assert.ThrowsAsync<ApiException>(() => client.GetFileBehaviorSummaryAsync("abc"));
-    }
-
-    [Fact]
-    public async Task GetFileNetworkTrafficAsync_DeserializesResponse()
-    {
-        var json = "{\"data\":{\"tcp\":[{\"dst\":\"1.2.3.4\",\"port\":80}]}}";
-        var handler = new SingleResponseHandler(new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(json, Encoding.UTF8, "application/json")
-        });
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        var traffic = await client.GetFileNetworkTrafficAsync("abc");
-
-        Assert.NotNull(traffic);
-        Assert.NotNull(handler.Request);
-        Assert.Equal("/api/v3/files/abc/network-traffic", handler.Request!.RequestUri!.AbsolutePath);
-        Assert.Equal("1.2.3.4", traffic!.Data.Tcp[0].Destination);
-        Assert.Equal(80, traffic.Data.Tcp[0].Port);
-    }
-
-    [Fact]
-    public async Task GetFileNetworkTrafficAsync_ThrowsApiException()
-    {
-        var errorJson = @"{""error"":{""code"":""NotFoundError"",""message"":""not found""}}";
-        var response = new HttpResponseMessage(HttpStatusCode.NotFound)
-        {
-            Content = new StringContent(errorJson, Encoding.UTF8, "application/json")
-        };
-        var handler = new SingleResponseHandler(response);
-        var httpClient = new HttpClient(handler)
-        {
-            BaseAddress = new Uri("https://www.virustotal.com/api/v3/")
-        };
-        var client = new VirusTotalClient(httpClient);
-
-        await Assert.ThrowsAsync<ApiException>(() => client.GetFileNetworkTrafficAsync("abc"));
-    }
 }
